Add files via upload
I was trying to model different modalities with different FCN before concatenating them and sending them to a shared FCN. This approach was taking too long to train and was abandoned. 
We also learnt that because I was flattening the matrix in 1d vector before turning it into 2d matrix, we were loosing spatial dependencies.